Moving create_indexes to copy_table_metadata
